src/pyproject.toml
[project]
name = "moneyroundup"
version = "0.1.0"
description = "FastAPI backend for moneyroundup application"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Sosa Edison", email = "me@sosaedison.io" },
]
dependencies = [
    "fastapi>=0.95.2",
    "uvicorn>=0.22.0",
    "gunicorn>=20.1.0",
    "starlette>=0.27.0",
    "pydantic>=1.10.8,<2",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "plaid-python>=14.0.0",
    "supabase>=2.0.0",
    "PyJWT[crypto]>=2.6.0",
    "APScheduler>=3.10.1",
    "boto3>=1.26.0",
    "httpx>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
